+++ trunk/src/com/letlink/common/datamodel/Imageinfo.java
@@ -5,19 +5,19 @@ import java.io.Serializable;
 public class Imageinfo implements Serializable {
 
 	private static final long serialVersionUID = 2554670667148601075L;
-	private int imageinfoID;
+	private long imageinfoID;
 	private String filename;
 	private String fileFormat;
 	private String imageURL;
 	private String pageURL;
-	private int pixelsWidth;
-	private int pixelsHeight;
-	private int sizeBytes;
+	private long pixelsWidth;
+	private long pixelsHeight;
+	private long sizeBytes;
 	
 	public Imageinfo(){}
 
 	public Imageinfo(String filename, String fileFormat, String imageURL,
-			String pageURL, int pixelsWidth, int pixelsHeight, int sizeBytes) {
+			String pageURL, int pixelsWidth, int pixelsHeight, long sizeBytes) {
 		super();
 		this.filename = filename;
 		this.fileFormat = fileFormat;
@@ -28,11 +28,11 @@ public class Imageinfo implements Serializable {
 		this.sizeBytes = sizeBytes;
 	}
 
-	public int getImageinfoID() {
+	public long getImageinfoID() {
 		return imageinfoID;
 	}
 
-	public void setImageinfoID(int imageinfoID) {
+	public void setImageinfoID(long imageinfoID) {
 		this.imageinfoID = imageinfoID;
 	}
 
@@ -68,27 +68,27 @@ public class Imageinfo implements Serializable {
 		this.pageURL = pageURL;
 	}
 
-	public int getPixelsWidth() {
+	public long getPixelsWidth() {
 		return pixelsWidth;
 	}
 
-	public void setPixelsWidth(int pixelsWidth) {
+	public void setPixelsWidth(long pixelsWidth) {
 		this.pixelsWidth = pixelsWidth;
 	}
 
-	public int getPixelsHeight() {
+	public long getPixelsHeight() {
 		return pixelsHeight;
 	}
 
-	public void setPixelsHeight(int pixelsHeight) {
+	public void setPixelsHeight(long pixelsHeight) {
 		this.pixelsHeight = pixelsHeight;
 	}
 
-	public int getSizeBytes() {
+	public long getSizeBytes() {
 		return sizeBytes;
 	}
 
-	public void setSizeBytes(int sizeBytes) {
+	public void setSizeBytes(long sizeBytes) {
 		this.sizeBytes = sizeBytes;
 	}
